added types, rename settings_controller
