+++ backend/src/api/controllers/__init__.py
@@ -1,6 +1,6 @@
 from backend.src.api.controllers.record_controller import RecordController
 from backend.src.api.controllers.report_controller import ReportController
-from backend.src.api.controllers.settings_controller import SettingsController
+from backend.src.api.controllers.profile_controller import SettingsController
 from backend.src.api.controllers.task_controller import TaskController
 from backend.src.api.controllers.auth_controller import AuthController
