+++ frontend/src/lib/types/AIProviders.ts
@@ -1,0 +1,5 @@
+export interface TaskType {
+    id: number
+    name: string
+    is_active: boolean
+}
