+++ frontend/src/lib/types/settings.ts
@@ -1,3 +1,4 @@
 export interface TaskType {
     title: string
+    color: string
 }
