+++ frontend/src/lib/types/user.ts
@@ -1,4 +1,12 @@
 export interface User {
+    username: string
     email: string
     password: string
+    display_name: string
+    department: string
+    position: string
+    ai_auto_process: boolean
+    ai_provider_id: number
+    is_active: boolean
+    is_verify: boolean
 }
